requirements.txt
# ── DSPy Medical AI System — Dependencies ──

# Core AI Framework
dspy>=2.5.0
openai>=1.40.0

# Embeddings & Vector Store
langchain>=0.3.26
langchain-pinecone>=0.2.8
langchain-openai>=0.3.24
langchain-community>=0.3.26
sentence-transformers>=4.1.0
pinecone-client>=5.0.0

# Web Framework (FastAPI replaces Flask)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
jinja2>=3.1.4

# Data Validation & Settings
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Document Processing
pypdf>=5.6.1

# Environment & Config
python-dotenv>=1.1.0

# Multi-Agent Collaboration (AutoGen)
pyautogen>=0.4.0

# Observability
structlog>=24.0.0

# Package
-e .